third_party/blink/web_tests/http/tests/websocket/too-long-payload-expected.txt

Tests whether WebSocket correctly aborts the connection when it receives a frame with too long payload.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
